Updated separation of time.scala and main.scala
also updated any incorrect assumptions made
